Messages notify when something happen.

+++ src/containers/Alert.tsx
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { AlertAction, setAlert, setState, StateAction } from "../actions";
+import { addErrorMessage, addInfoMessage, AlertAction, MessagesAction, setAlert, setState, StateAction } from "../actions";
 import { deletePhotos, PhotoListAction } from "../actions/photoList";
 import { deleteVideos, VideoAction } from "../actions/videoList";
 import api from "../api";
@@ -17,7 +17,7 @@ function mapStateToProps(state: StoreState) {
   }
 }
 
-function mapDispatchToProps(dispatch: Dispatch<AlertAction | PhotoListAction | VideoAction | StateAction>) {
+function mapDispatchToProps(dispatch: Dispatch<AlertAction | PhotoListAction | VideoAction | StateAction | MessagesAction>) {
   return {
     onDelete: (names: string[]) => {
       dispatch(setAlert(AlertState.Deleting))
@@ -39,15 +39,24 @@ function mapDispatchToProps(dispatch: Dispatch<AlertAction | PhotoListAction | V
           }
         }
 
-        dispatch(deletePhotos(deletedPhotosMap))
-        dispatch(deleteVideos(deletedVideosMap))
+        if (deletedPhotosMap.size) {
+          dispatch(deletePhotos(deletedPhotosMap))
+          dispatch(addInfoMessage('刪除成功', `刪除了${deletedPhotosMap.size}張相片`))
+        }
+        if (deletedVideosMap.size) {
+          dispatch(deleteVideos(deletedVideosMap))
+          dispatch(addInfoMessage('刪除成功', `刪除了${deletedVideosMap.size}部影片`))
+        }
       })
-      .catch(err => console.log(err))
+      .catch(err => {
+        console.log(err)
+        dispatch(addErrorMessage('刪除檔案時發生異常', err.response.data))
+      })
       .finally(() => {
         setTimeout(() => {
           dispatch(setAlert(AlertState.None))
           dispatch(setState(State.Browse))
-        }, 1500)
+        }, 500)
       })
     },
     onCancel: () => dispatch(setAlert(AlertState.None))
